Bug fix for super scaffolding
Currently it's not possible to generate an audit log. This PR along with
this one from `core`
(https://github.com/bullet-train-co/bullet_train-core/pull/945) fix that
particulart problem.

This is really just a quick and dirty fix to get things working as
quickly as possible. Eventually we should remove this method and use
whatever the new standard method is.